Countdown rework

+++ Source/ShootThemUp/Private/UI/STUSpectatorWidget.cpp
@@ -5,11 +5,11 @@
 
 #include "STURespawnComponent.h"
 
-bool USTUSpectatorWidget::GetRespawnTime(int32& CountDownTime) const
+bool USTUSpectatorWidget::GetRespawnTime(float& CountDownTime) const
 {
 	const APlayerController* PlayerController = GetOwningPlayer();
 	if (!PlayerController) return false;
-	
+
 	const auto RespawnComponent = PlayerController->FindComponentByClass<USTURespawnComponent>();
 	if (!RespawnComponent || !RespawnComponent->IsRespawnInProgress()) return false;
